fix #3: improve the way it search for the solution directory
Note:

It worked fine when using `dotnet run` directly in the Demo folder or when using Rider.
But in Visual Studio, `Directory.GetCurrentDirectory()` give us the binary folder (with the exe) and not the project folder as I were expected.
The fix here is to start from the `Directory.GetCurrentDirectory()` and search in parents for the .sln file, which will be located in the solution directory.
I don't expect it to run outside of a IDE, but it seem acceptable here since it is just a demo.